Mutations now create new ind; file io changed; bugfixes (see below)
This should be an pretty good list of the changes:

x) Changed the mutations so that they create a new ID.

x) Changed the LAMMPS and FEMSIM code to prevent saving/copying duplicate
files. This should save on file IO time as well as disk space.

x) All model file data are saved to loggingdir/modelfiles

x) Completely rewrote crossmodule/lammps.py to enable specifying of the
output trajectory file.

x) Update the parameters file with "keep_best" as a parameter to the
predator.

x) Removed the mutations parameter "preserve_best" because
"keep_original_best" is better functionality.

x) If a LAMMPS calculation fails, the individual's energy is set to
np.inf (rather than 0 or np.nan).

x) The `clear_XYZs` functionality no longer works because the
individuals are not saved in a specific generation's folder, they are
all saved at the same level in the `modelfiles` directory.

x) Fixed a couple bugs in Individual.copy()

x) Changed how the post processing parameters are used in genetic.py

x) Added a bit of post-processing to the predators.

x) I changed the format of the mutation and crossover tags. It should be
pretty self explanatory.

x) Mutated individuals now have a "mutated_from" attribute specifying
the ID of the individual from where they were mutated.